backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.40
asyncpg>=0.29.0
alembic>=1.13.0
structlog>=24.0.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
openpyxl>=3.1.0
